+++ src/types.ts
@@ -34,6 +34,7 @@ export interface SankeyChartConfig extends LovelaceCardConfig {
     print_yaml?: boolean;
     group_by_floor?: boolean;
     group_by_area?: boolean;
+    net_flows?: boolean;
   };
   title?: string;
   convert_units_to?: '' | CONVERSION_UNITS;